Add files via upload
Tria Sistema Operatiu 2.0.4 BETA for Windows (Aaron Edition for Win NT 5.1 to 6.0 and Standard edition for Win NT 6.1 and newer ones) with the following news:

A) Detects and recommends Windows language in the 1st screen, keeping the possibility to use 14 different languages. If Windows is not in one of this 14 languages, recommends English.

B) Improved Operating System details, including if is HOME, PRO or other version, and Service Pack(s) (if there is one).

C) Info/FAQs screen updated and optimized, in 14 languages.